Feature/melhorias gerais nodes para clientes

See merge request getinsight.it/produtos/access-pilot/accesspilot-backend!4

+++ src/main/resources/query/find-all-requests-in-roles.sql
@@ -9,4 +9,4 @@ inner join TB_ROLE r on s.id_role = r.id
 where r.nome in (:roles)
   AND (:clientIds IS NULL OR r.id_cliente = ANY(:clientIds))
   AND (:levelIds  IS NULL OR s.esfera_id  = ANY(:levelIds))
-  AND (:itemIds   IS NULL OR s.codigo_item::bigint = ANY(:itemIds))
+  AND (:itemIds   IS NULL OR s.codigo_item = ANY(:itemIds))
